Refactor project structure: rename base package of search-engine-uni-backend to de.uni_leipzig.search_engine_uni.backend
