+++ search-engine-uni-backend/src/main/java/de/uni_leipzig/search_engine_uni/backend/aspects/EnrichResponsesWithViewAspect.java
@@ -1,4 +1,4 @@
-package de.uni_leipzig.search_engine_uni.aspects;
+package de.uni_leipzig.search_engine_uni.backend.aspects;
 
 import java.util.Arrays;
 import java.util.Collections;
