+++ search-engine-uni-backend/src/main/java/de/uni_leipzig/search_engine_uni/backend/aspects/dto/RequestDto.java
@@ -1,4 +1,4 @@
-package de.uni_leipzig.search_engine_uni.aspects.dto;
+package de.uni_leipzig.search_engine_uni.backend.aspects.dto;
 
 import java.util.Enumeration;
 import java.util.HashMap;
