+++ search-engine-uni-backend/src/main/java/de/uni_leipzig/search_engine_uni/backend/controller/RedirectController.java
@@ -1,4 +1,4 @@
-package de.uni_leipzig.search_engine_uni.controller;
+package de.uni_leipzig.search_engine_uni.backend.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.view.RedirectView;
 
-import de.uni_leipzig.search_engine_uni.dto.SearchResult;
-import de.uni_leipzig.search_engine_uni.lucene.SearcherComponent;
+import de.uni_leipzig.search_engine_uni.backend.dto.SearchResult;
+import de.uni_leipzig.search_engine_uni.backend.lucene.SearcherComponent;
 
 @Controller
 public class RedirectController
