+++ search-engine-uni-backend/src/main/java/de/uni_leipzig/search_engine_uni/backend/dto/SearchResult.java
@@ -1,11 +1,11 @@
-package de.uni_leipzig.search_engine_uni.dto;
+package de.uni_leipzig.search_engine_uni.backend.dto;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.lucene.document.Document;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.mvc.ControllerLinkBuilder;
 
-import de.uni_leipzig.search_engine_uni.controller.RedirectController;
+import de.uni_leipzig.search_engine_uni.backend.controller.RedirectController;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
