+++ search-engine-uni-backend/src/test/java/de/uni_leipzig/search_engine_uni/backend/controller/SearchControllerLinksTest.checkEmptyResultAsInput.approved.txt
