+++ search-engine-uni-backend/src/test/java/de/uni_leipzig/search_engine_uni/backend/controller/SearchControllerTest.approveNullInput.approved.txt
